minor fixes
changed notification model to have created_at field

Upon changing the model with existing data:
python3 manage.py makemigrations
It is impossible to add the field 'created_at' with 'auto_now_add=True' to notifications without providing a default. This is because the database needs something to populate existing rows.
So, can add timezone.now for the existing data in the options that it serves